+++ eras/allegra/impl/src/Cardano/Ledger/Allegra/Core.hs
@@ -6,5 +6,6 @@ module Cardano.Ledger.Allegra.Core (
 where
 
 import Cardano.Ledger.Allegra.Scripts (ValidityInterval (..))
+import Cardano.Ledger.Allegra.Tx ()
 import Cardano.Ledger.Allegra.TxBody (AllegraEraTxBody (..))
 import Cardano.Ledger.Shelley.Core