+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/Core.hs
@@ -31,6 +31,7 @@ import Cardano.Ledger.Core (
  )
 import Cardano.Ledger.Core hiding (EraPParams (..))
 import Cardano.Ledger.Shelley.Governance
+import Cardano.Ledger.Shelley.Tx ()
 import Cardano.Ledger.Shelley.TxBody (ShelleyEraTxBody (..))
 import Cardano.Ledger.Shelley.TxCert (
   MIRCert (..),